chore: bulk migrate remaining session service interfaces to core
- Moved Config, Agent, Skill, Provider types to @opencode-ai/core
- Moved Permission full impl, Storage/Record/EffectHttpClient utils to core
- Updated 40+ session files to import from @opencode-ai/core
- Remaining @/-imports are runtime-only (defaultLayer, node)
