+++ packages/opencode/src/server/routes/instance/httpapi/groups/instance.ts
@@ -1,9 +1,9 @@
-import { Agent } from "@/agent/agent"
+import { Agent } from "@opencode-ai/core/agent/agent"
 import { Command } from "@/command"
 import { Format } from "@/format"
 import { LSP } from "@/lsp/lsp"
 import { Vcs } from "@/project/vcs"
-import { Skill } from "@/skill"
+import { Skill } from "@opencode-ai/core/skill/index"
 import { Schema } from "effect"
 import { HttpApi, HttpApiEndpoint, HttpApiGroup, HttpApiSchema, OpenApi } from "effect/unstable/httpapi"
 import { Authorization } from "../middleware/authorization"
