+++ packages/opencode/src/server/routes/instance/httpapi/handlers/project-copy.ts
@@ -1,4 +1,4 @@
-import { Agent } from "@/agent/agent"
+import { Agent } from "@opencode-ai/core/agent/agent"
 import { Provider } from "@/provider/provider"
 import { LLM } from "@/session/llm"
 import { MessageID, SessionID } from "@/session/schema"
