+++ packages/opencode/src/session/llm/native-request.ts
@@ -10,8 +10,8 @@ import {
   OpenRouter,
 } from "@opencode-ai/llm/providers"
 import type { ModelMessage } from "ai"
-import type { Provider } from "@/provider/provider"
-import { isRecord } from "@/util/record"
+import type { Provider } from "@opencode-ai/core/provider/provider"
+import { isRecord } from "@opencode-ai/core/util/record"
 
 type ToolInput = {
   readonly description?: string
