+++ packages/opencode/src/session/llm/native-runtime.ts
@@ -1,8 +1,8 @@
 import type { Auth } from "@/auth"
-import type { Provider } from "@/provider/provider"
-import { ProviderTransform } from "@/provider/transform"
+import type { Provider } from "@opencode-ai/core/provider/provider"
+import { ProviderTransform } from "@opencode-ai/core/provider/transform"
 import { errorMessage } from "@/util/error"
-import { isRecord } from "@/util/record"
+import { isRecord } from "@opencode-ai/core/util/record"
 import { asSchema, type ModelMessage, type Tool } from "ai"
 import { Cause, Effect, FiberSet, Queue } from "effect"
 import * as Stream from "effect/Stream"
