+++ packages/opencode/src/session/llm/request.ts
@@ -1,13 +1,13 @@
 import { PermissionV1 } from "@opencode-ai/core/v1/permission"
 import type { Auth } from "@/auth"
 import { SessionV1 } from "@opencode-ai/core/v1/session"
-import type { RuntimeFlags } from "@/effect/runtime-flags"
-import { InstanceState } from "@/effect/instance-state"
-import { Permission } from "@/permission"
-import type { Agent } from "@/agent/agent"
+import type { RuntimeFlags } from "@opencode-ai/core/effect/runtime-flags"
+import { InstanceState } from "@opencode-ai/core/effect/instance-state"
+import { Permission } from "@opencode-ai/core/permission"
+import type { Agent } from "@opencode-ai/core/agent/agent"
 import type { MessageV2 } from "../message-v2"
-import type { Provider } from "@/provider/provider"
-import { ProviderTransform } from "@/provider/transform"
+import type { Provider } from "@opencode-ai/core/provider/provider"
+import { ProviderTransform } from "@opencode-ai/core/provider/transform"
 import { SystemPrompt } from "../system"
 import { InstallationVersion } from "@opencode-ai/core/installation/version"
 import { Effect, Record } from "effect"
