+++ packages/opencode/src/session/overflow.ts
@@ -1,8 +1,7 @@
-import type { Config } from "@/config/config"
 import { ConfigV1 } from "@opencode-ai/core/v1/config/config"
 import { SessionV1 } from "@opencode-ai/core/v1/session"
-import type { Provider } from "@/provider/provider"
-import { ProviderTransform } from "@/provider/transform"
+import type { Provider } from "@opencode-ai/core/provider/provider"
+import { ProviderTransform } from "@opencode-ai/core/provider/transform"
 import type { MessageV2 } from "./message-v2"
 
 const COMPACTION_BUFFER = 20_000
