+++ packages/opencode/src/session/reminders.ts
@@ -1,10 +1,10 @@
 import path from "path"
 import { SessionV1 } from "@opencode-ai/core/v1/session"
 import { Effect } from "effect"
-import { Agent } from "@/agent/agent"
+import { Agent } from "@opencode-ai/core/agent/agent"
 import { FSUtil } from "@opencode-ai/core/fs-util"
-import { InstanceState } from "@/effect/instance-state"
-import { RuntimeFlags } from "@/effect/runtime-flags"
+import { InstanceState } from "@opencode-ai/core/effect/instance-state"
+import { RuntimeFlags } from "@opencode-ai/core/effect/runtime-flags"
 import { PartID } from "./schema"
 import { MessageV2 } from "./message-v2"
 import { Session } from "./session"
