+++ packages/opencode/src/session/retry.ts
@@ -2,8 +2,8 @@ import type { NamedError } from "@opencode-ai/core/util/error"
 import { SessionV1 } from "@opencode-ai/core/v1/session"
 import { Cause, Clock, Duration, Effect, Schedule } from "effect"
 import { MessageV2 } from "./message-v2"
-import { iife } from "@/util/iife"
-import { isRecord } from "@/util/record"
+import { iife } from "@opencode-ai/core/util/iife"
+import { isRecord } from "@opencode-ai/core/util/record"
 
 export type Err = ReturnType<NamedError["toObject"]>
