+++ packages/opencode/src/session/revert.ts
@@ -1,9 +1,10 @@
 import { LayerNode } from "@opencode-ai/core/effect/layer-node"
 import { Effect, Layer, Context, Schema } from "effect"
 import { SessionV1 } from "@opencode-ai/core/v1/session"
-import { EventV2Bridge } from "@/event-v2-bridge"
+import { EventV2Bridge } from "@opencode-ai/core/event-v2-bridge"
 import { Snapshot } from "../snapshot"
-import { Storage } from "@/storage/storage"
+import type { Storage } from "@opencode-ai/core/storage/storage"
+import { defaultLayer as _stLayer, node as _stNode } from "@/storage/storage"
 import { Session } from "./session"
 import { MessageV2 } from "./message-v2"
 import { SessionID, MessageID, PartID } from "./schema"
