+++ packages/opencode/src/session/tools.ts
@@ -1,9 +1,9 @@
-import { Agent } from "@/agent/agent"
+import { Agent } from "@opencode-ai/core/agent/agent"
 import { SessionV1 } from "@opencode-ai/core/v1/session"
-import { Provider } from "@/provider/provider"
-import { ProviderTransform } from "@/provider/transform"
+import { Provider } from "@opencode-ai/core/provider/provider"
+import { ProviderTransform } from "@opencode-ai/core/provider/transform"
 import { MCP } from "@/mcp"
-import { Permission } from "@/permission"
+import { Permission } from "@opencode-ai/core/permission"
 import { Tool } from "@/tool/tool"
 import { ToolJsonSchema } from "@/tool/json-schema"
 import { ToolRegistry } from "@/tool/registry"
