+++ packages/opencode/src/tool/task.ts
@@ -6,7 +6,7 @@ import { BackgroundJob } from "@/background/job"
 import { Session } from "@/session/session"
 import { SessionID, MessageID } from "../session/schema"
 import { MessageV2 } from "../session/message-v2"
-import { Agent } from "../agent/agent"
+import { Agent } from "@opencode-ai/core/agent/agent"
 import { deriveSubagentSessionPermission } from "../agent/subagent-permissions"
 import type { SessionPrompt } from "../session/prompt"
 import { Config } from "@/config/config"
